feat(member): 貼 /shop/c 連結到 LINE 直接出現商品縮圖與團名
貼到 LINE / FB 的連結預覽讀的是伺服器吐出來的 og tag，而 /shop/c/[id] 整頁
是 client component，爬蟲抓不到任何資料 —— 全站的團連結預覽都長一樣：站台
標題 + LINE 退回 apple-touch-icon 的店家 logo 小方圖。

- page.tsx 改成 server wrapper（原本的內容搬到 CampaignDetailClient.tsx，
  行為不動），只多一支 generateMetadata：標題＝團名、圖＝團封面（沒設就退回
  第一個品項的商品主圖，再沒有才退店家 banner）、說明＝「$N 起・文案前 90 字」。
- liff-api 新增免 token action get_campaign_preview：爬蟲沒有會員 JWT，
  掛在需要 token 的區塊等於整支失效。只回團名 / 圖 / 起跳價 / 結單時間。
  非 is_for_shop 與內部補貨 sentinel 擋掉；已結單的團照樣回（連結已經發出去了，
  顯示團名與圖比通用卡有用，頁面本身仍會擋下單）。
- root layout 補上站台預設 openGraph，其餘頁面（/shop、/spot…）至少是
  店家 banner + 店名，不再是那張小方圖。
- SITE_URL / SITE_NAME / banner 收進 lib/site.ts，/join 改用同一份。

頁面外殼對所有人都一樣（資料仍由 client 端登入後抓），所以整頁連同 metadata
ISR 快取 300 秒，不用每次瀏覽都多打一次 edge function。
