+++ apps/member/public/sw.js
@@ -1,2 +1,2 @@
 (()=>{"use strict";let e,t,a,s,r,n={googleAnalytics:"googleAnalytics",precache:"precache-v2",prefix:"serwist",runtime:"runtime",suffix:"u">typeof registration?registration.scope:""},i=e=>[n.prefix,e,n.suffix].filter(e=>e&&e.length>0).join("-"),c=e=>e||i(n.precache),o=e=>e||i(n.runtime);var l=class extends Error{details;constructor(e,t){super(((e,...t)=>{let a=e;return t.length>0&&(a+=` :: ${JSON.stringify(t)}`),a})(e,t)),this.name=e,this.details=t}};function h(e){return new Promise(t=>setTimeout(t,e))}let u=new Set;function d(e,t){let a=new URL(e);for(let e of t)a.searchParams.delete(e);return a.href}async function m(e,t,a,s){let r=d(t.url,a);if(t.url===r)return e.match(t,s);let n={...s,ignoreSearch:!0};for(let i of(await e.keys(t,n)))if(r===d(i.url,a))return e.match(i,s)}var f=class{promise;resolve;reject;constructor(){this.promise=new Promise((e,t)=>{this.resolve=e,this.reject=t})}};let g=async()=>{for(let e of u)await e()},w="-precache-",p=async(e,t=w)=>{let a=(await self.caches.keys()).filter(a=>a.includes(t)&&a.includes(self.registration.scope)&&a!==e);return await Promise.all(a.map(e=>self.caches.delete(e))),a},y=(e,t)=>{let a=t();return e.waitUntil(a),a},_=(e,t)=>t.some(t=>e instanceof t),b=new WeakMap,x=new WeakMap,v=new WeakMap,E={get(e,t,a){if(e instanceof IDBTransaction){if("done"===t)return b.get(e);if("store"===t)return a.objectStoreNames[1]?void 0:a.objectStore(a.objectStoreNames[0])}return R(e[t])},set:(e,t,a)=>(e[t]=a,!0),has:(e,t)=>e instanceof IDBTransaction&&("done"===t||"store"===t)||t in e};function R(e){if(e instanceof IDBRequest){let t;return t=new Promise((t,a)=>{let s=()=>{e.removeEventListener("success",r),e.removeEventListener("error",n)},r=()=>{t(R(e.result)),s()},n=()=>{a(e.error),s()};e.addEventListener("success",r),e.addEventListener("error",n)}),v.set(t,e),t}if(x.has(e))return x.get(e);let t=function(e){if("function"==typeof e)return(r||(r=[IDBCursor.prototype.advance,IDBCursor.prototype.continue,IDBCursor.prototype.continuePrimaryKey])).includes(e)?function(...t){return e.apply(q(this),t),R(this.request)}:function(...t){return R(e.apply(q(this),t))};return(e instanceof IDBTransaction&&function(e){if(b.has(e))return;let t=new Promise((t,a)=>{let s=()=>{e.removeEventListener("complete",r),e.removeEventListener("error",n),e.removeEventListener("abort",n)},r=()=>{t(),s()},n=()=>{a(e.error||new DOMException("AbortError","AbortError")),s()};e.addEventListener("complete",r),e.addEventListener("error",n),e.addEventListener("abort",n)});b.set(e,t)}(e),_(e,s||(s=[IDBDatabase,IDBObjectStore,IDBIndex,IDBCursor,IDBTransaction])))?new Proxy(e,E):e}(e);return t!==e&&(x.set(e,t),v.set(t,e)),t}let q=e=>v.get(e);function S(e,t,{blocked:a,upgrade:s,blocking:r,terminated:n}={}){let i=indexedDB.open(e,t),c=R(i);return s&&i.addEventListener("upgradeneeded",e=>{s(R(i.result),e.oldVersion,e.newVersion,R(i.transaction),e)}),a&&i.addEventListener("blocked",e=>a(e.oldVersion,e.newVersion,e)),c.then(e=>{n&&e.addEventListener("close",()=>n()),r&&e.addEventListener("versionchange",e=>r(e.oldVersion,e.newVersion,e))}).catch(()=>{}),c}let D=["get","getKey","getAll","getAllKeys","count"],N=["put","add","delete","clear"],C=new Map;function T(e,t){if(!(e instanceof IDBDatabase&&!(t in e)&&"string"==typeof t))return;if(C.get(t))return C.get(t);let a=t.replace(/FromIndex$/,""),s=t!==a,r=N.includes(a);if(!(a in(s?IDBIndex:IDBObjectStore).prototype)||!(r||D.includes(a)))return;let n=async function(e,...t){let n=this.transaction(e,r?"readwrite":"readonly"),i=n.store;return s&&(i=i.index(t.shift())),(await Promise.all([i[a](...t),r&&n.done]))[0]};return C.set(t,n),n}E={...e=E,get:(t,a,s)=>T(t,a)||e.get(t,a,s),has:(t,a)=>!!T(t,a)||e.has(t,a)};let P=["continue","continuePrimaryKey","advance"],A={},k=new WeakMap,I=new WeakMap,U={get(e,t){if(!P.includes(t))return e[t];let a=A[t];return a||(a=A[t]=function(...e){k.set(this,I.get(this)[t](...e))}),a}};async function*L(...e){let t=this;if(t instanceof IDBCursor||(t=await t.openCursor(...e)),!t)return;let a=new Proxy(t,U);for(I.set(a,t),v.set(a,q(t));t;)yield a,t=await (k.get(a)||t.continue()),k.delete(a)}function F(e,t){return t===Symbol.asyncIterator&&_(e,[IDBIndex,IDBObjectStore,IDBCursor])||"iterate"===t&&_(e,[IDBIndex,IDBObjectStore])}E={...t=E,get:(e,a,s)=>F(e,a)?L:t.get(e,a,s),has:(e,a)=>F(e,a)||t.has(e,a)};let M=async(e,t)=>{let s=null;if(e.url&&(s=new URL(e.url).origin),s!==self.location.origin)throw new l("cross-origin-copy-response",{origin:s});let r=e.clone(),n={headers:new Headers(r.headers),status:r.status,statusText:r.statusText},i=t?t(n):n,c=!function(){if(void 0===a){let e=new Response("");if("body"in e)try{new Response(e.body),a=!0}catch{a=!1}a=!1}return a}()?await r.blob():r.body;return new Response(c,i)},B="requests",O="queueName";var K=class{_db=null;async addEntry(e){let t=(await this.getDb()).transaction(B,"readwrite",{durability:"relaxed"});await t.store.add(e),await t.done}async getFirstEntryId(){return(await (await this.getDb()).transaction(B).store.openCursor())?.value.id}async getAllEntriesByQueueName(e){return await (await this.getDb()).getAllFromIndex(B,O,IDBKeyRange.only(e))||[]}async getEntryCountByQueueName(e){return(await this.getDb()).countFromIndex(B,O,IDBKeyRange.only(e))}async deleteEntry(e){await (await this.getDb()).delete(B,e)}async getFirstEntryByQueueName(e){return await this.getEndEntryFromIndex(IDBKeyRange.only(e),"next")}async getLastEntryByQueueName(e){return await this.getEndEntryFromIndex(IDBKeyRange.only(e),"prev")}async getEndEntryFromIndex(e,t){return(await (await this.getDb()).transaction(B).store.index(O).openCursor(e,t))?.value}async getDb(){return this._db||(this._db=await S("serwist-background-sync",3,{upgrade:this._upgradeDb})),this._db}_upgradeDb(e,t){t>0&&t<3&&e.objectStoreNames.contains(B)&&e.deleteObjectStore(B),e.createObjectStore(B,{autoIncrement:!0,keyPath:"id"}).createIndex(O,O,{unique:!1})}},W=class{_queueName;_queueDb;constructor(e){this._queueName=e,this._queueDb=new K}async pushEntry(e){delete e.id,e.queueName=this._queueName,await this._queueDb.addEntry(e)}async unshiftEntry(e){let t=await this._queueDb.getFirstEntryId();t?e.id=t-1:delete e.id,e.queueName=this._queueName,await this._queueDb.addEntry(e)}async popEntry(){return this._removeEntry(await this._queueDb.getLastEntryByQueueName(this._queueName))}async shiftEntry(){return this._removeEntry(await this._queueDb.getFirstEntryByQueueName(this._queueName))}async getAll(){return await this._queueDb.getAllEntriesByQueueName(this._queueName)}async size(){return await this._queueDb.getEntryCountByQueueName(this._queueName)}async deleteEntry(e){await this._queueDb.deleteEntry(e)}async _removeEntry(e){return e&&await this.deleteEntry(e.id),e}};let j=["method","referrer","referrerPolicy","mode","credentials","cache","redirect","integrity","keepalive"];var $=class e{_requestData;static async fromRequest(t){let a={url:t.url,headers:{}};for(let e of("GET"!==t.method&&(a.body=await t.clone().arrayBuffer()),t.headers.forEach((e,t)=>{a.headers[t]=e}),j))void 0!==t[e]&&(a[e]=t[e]);return new e(a)}constructor(e){"navigate"===e.mode&&(e.mode="same-origin"),this._requestData=e}toObject(){let e=Object.assign({},this._requestData);return e.headers=Object.assign({},this._requestData.headers),e.body&&(e.body=e.body.slice(0)),e}toRequest(){return new Request(this._requestData.url,this._requestData)}clone(){return new e(this.toObject())}};let H="serwist-background-sync",G=new Set,Q=e=>{let t={request:new $(e.requestData).toRequest(),timestamp:e.timestamp};return e.metadata&&(t.metadata=e.metadata),t};var V=class{_name;_onSync;_maxRetentionTime;_queueStore;_forceSyncFallback;_syncInProgress=!1;_requestsAddedDuringSync=!1;constructor(e,{forceSyncFallback:t,onSync:a,maxRetentionTime:s}={}){if(G.has(e))throw new l("duplicate-queue-name",{name:e});G.add(e),this._name=e,this._onSync=a||this.replayRequests,this._maxRetentionTime=s||10080,this._forceSyncFallback=!!t,this._queueStore=new W(this._name),this._addSyncListener()}get name(){return this._name}async pushRequest(e){await this._addRequest(e,"push")}async unshiftRequest(e){await this._addRequest(e,"unshift")}async popRequest(){return this._removeRequest("pop")}async shiftRequest(){return this._removeRequest("shift")}async getAll(){let e=await this._queueStore.getAll(),t=Date.now(),a=[];for(let s of e){let e=60*this._maxRetentionTime*1e3;t-s.timestamp>e?await this._queueStore.deleteEntry(s.id):a.push(Q(s))}return a}async size(){return await this._queueStore.size()}async _addRequest({request:e,metadata:t,timestamp:a=Date.now()},s){let r={requestData:(await $.fromRequest(e.clone())).toObject(),timestamp:a};switch(t&&(r.metadata=t),s){case"push":await this._queueStore.pushEntry(r);break;case"unshift":await this._queueStore.unshiftEntry(r)}this._syncInProgress?this._requestsAddedDuringSync=!0:await this.registerSync()}async _removeRequest(e){let t,a=Date.now();switch(e){case"pop":t=await this._queueStore.popEntry();break;case"shift":t=await this._queueStore.shiftEntry()}if(t){let s=60*this._maxRetentionTime*1e3;return a-t.timestamp>s?this._removeRequest(e):Q(t)}}async replayRequests(){let e;for(;e=await this.shiftRequest();)try{await fetch(e.request.clone())}catch{throw await this.unshiftRequest(e),new l("queue-replay-failed",{name:this._name})}}async registerSync(){if("sync"in self.registration&&!this._forceSyncFallback)try{await self.registration.sync.register(`${H}:${this._name}`)}catch(e){}}_addSyncListener(){"sync"in self.registration&&!this._forceSyncFallback?self.addEventListener("sync",e=>{if(e.tag===`${H}:${this._name}`){let t=async()=>{let t;this._syncInProgress=!0;try{await this._onSync({queue:this})}catch(e){if(e instanceof Error)throw e}finally{this._requestsAddedDuringSync&&!(t&&!e.lastChance)&&await this.registerSync(),this._syncInProgress=!1,this._requestsAddedDuringSync=!1}};e.waitUntil(t())}}):this._onSync({queue:this})}static get _queueNames(){return G}},z=class{_queue;constructor(e,t){this._queue=new V(e,t)}async fetchDidFail({request:e}){await this._queue.pushRequest({request:e})}};let J={cacheWillUpdate:async({response:e})=>200===e.status||0===e.status?e:null};function X(e){return"string"==typeof e?new Request(e):e}var Y=class{event;request;url;params;_cacheKeys={};_strategy;_handlerDeferred;_extendLifetimePromises;_plugins;_pluginStateMap;constructor(e,t){for(const a of(this.event=t.event,this.request=t.request,t.url&&(this.url=t.url,this.params=t.params),this._strategy=e,this._handlerDeferred=new f,this._extendLifetimePromises=[],this._plugins=[...e.plugins],this._pluginStateMap=new Map,this._plugins))this._pluginStateMap.set(a,{});this.event.waitUntil(this._handlerDeferred.promise)}async fetch(e){let{event:t}=this,a=X(e),s=await this.getPreloadResponse();if(s)return s;let r=this.hasCallback("fetchDidFail")?a.clone():null;try{for(let e of this.iterateCallbacks("requestWillFetch"))a=await e({request:a.clone(),event:t})}catch(e){if(e instanceof Error)throw new l("plugin-error-request-will-fetch",{thrownErrorMessage:e.message})}let n=a.clone();try{let e;for(let s of(e=await fetch(a,"navigate"===a.mode?void 0:this._strategy.fetchOptions),this.iterateCallbacks("fetchDidSucceed")))e=await s({event:t,request:n,response:e});return e}catch(e){throw r&&await this.runCallbacks("fetchDidFail",{error:e,event:t,originalRequest:r.clone(),request:n.clone()}),e}}async fetchAndCachePut(e){let t=await this.fetch(e),a=t.clone();return this.waitUntil(this.cachePut(e,a)),t}async cacheMatch(e){let t,a=X(e),{cacheName:s,matchOptions:r}=this._strategy,n=await this.getCacheKey(a,"read"),i={...r,cacheName:s};for(let e of(t=await caches.match(n,i),this.iterateCallbacks("cachedResponseWillBeUsed")))t=await e({cacheName:s,matchOptions:r,cachedResponse:t,request:n,event:this.event})||void 0;return t}async cachePut(e,t){let a=X(e);await h(0);let s=await this.getCacheKey(a,"write");if(!t)throw new l("cache-put-with-no-response",{url:new URL(String(s.url),location.href).href.replace(RegExp(`^${location.origin}`),"")});let r=await this._ensureResponseSafeToCache(t);if(!r)return!1;let{cacheName:n,matchOptions:i}=this._strategy,c=await self.caches.open(n),o=this.hasCallback("cacheDidUpdate"),u=o?await m(c,s.clone(),["__WB_REVISION__"],i):null;try{await c.put(s,o?r.clone():r)}catch(e){if(e instanceof Error)throw"QuotaExceededError"===e.name&&await g(),e}for(let e of this.iterateCallbacks("cacheDidUpdate"))await e({cacheName:n,oldResponse:u,newResponse:r.clone(),request:s,event:this.event});return!0}async getCacheKey(e,t){let a=`${e.url} | ${t}`;if(!this._cacheKeys[a]){let s=e;for(let e of this.iterateCallbacks("cacheKeyWillBeUsed"))s=X(await e({mode:t,request:s,event:this.event,params:this.params}));this._cacheKeys[a]=s}return this._cacheKeys[a]}hasCallback(e){for(let t of this._strategy.plugins)if(e in t)return!0;return!1}async runCallbacks(e,t){for(let a of this.iterateCallbacks(e))await a(t)}*iterateCallbacks(e){for(let t of this._strategy.plugins)if("function"==typeof t[e]){let a=this._pluginStateMap.get(t),s=s=>{let r={...s,state:a};return t[e](r)};yield s}}waitUntil(e){return this._extendLifetimePromises.push(e),e}async doneWaiting(){let e;for(;e=this._extendLifetimePromises.shift();)await e}destroy(){this._handlerDeferred.resolve(null)}async getPreloadResponse(){if(this.event instanceof FetchEvent&&"navigate"===this.event.request.mode&&"preloadResponse"in this.event)try{let e=await this.event.preloadResponse;if(e)return e}catch(e){return}}async _ensureResponseSafeToCache(e){let t=e,a=!1;for(let e of this.iterateCallbacks("cacheWillUpdate"))if(t=await e({request:this.request,response:t,event:this.event})||void 0,a=!0,!t)break;return!a&&t&&200!==t.status&&(t=void 0),t}},Z=class{cacheName;plugins;fetchOptions;matchOptions;constructor(e={}){this.cacheName=o(e.cacheName),this.plugins=e.plugins||[],this.fetchOptions=e.fetchOptions,this.matchOptions=e.matchOptions}handle(e){let[t]=this.handleAll(e);return t}handleAll(e){e instanceof FetchEvent&&(e={event:e,request:e.request});let t=e.event,a="string"==typeof e.request?new Request(e.request):e.request,s=new Y(this,e.url?{event:t,request:a,url:e.url,params:e.params}:{event:t,request:a}),r=this._getResponse(s,a,t);return[r,this._awaitComplete(r,s,a,t)]}async _getResponse(e,t,a){let s;await e.runCallbacks("handlerWillStart",{event:a,request:t});try{if(s=await this._handle(t,e),void 0===s||"error"===s.type)throw new l("no-response",{url:t.url})}catch(r){if(r instanceof Error){for(let n of e.iterateCallbacks("handlerDidError"))if(void 0!==(s=await n({error:r,event:a,request:t})))break}if(!s)throw r}for(let r of e.iterateCallbacks("handlerWillRespond"))s=await r({event:a,request:t,response:s});return s}async _awaitComplete(e,t,a,s){let r,n;try{r=await e}catch{}try{await t.runCallbacks("handlerDidRespond",{event:s,request:a,response:r}),await t.doneWaiting()}catch(e){e instanceof Error&&(n=e)}if(await t.runCallbacks("handlerDidComplete",{event:s,request:a,response:r,error:n}),t.destroy(),n)throw n}},ee=class extends Z{_networkTimeoutSeconds;constructor(e={}){super(e),this.plugins.some(e=>"cacheWillUpdate"in e)||this.plugins.unshift(J),this._networkTimeoutSeconds=e.networkTimeoutSeconds||0}async _handle(e,t){let a,s=[],r=[];if(this._networkTimeoutSeconds){let{id:n,promise:i}=this._getTimeoutPromise({request:e,logs:s,handler:t});a=n,r.push(i)}let n=this._getNetworkPromise({timeoutId:a,request:e,logs:s,handler:t});r.push(n);let i=await t.waitUntil((async()=>await t.waitUntil(Promise.race(r))||await n)());if(!i)throw new l("no-response",{url:e.url});return i}_getTimeoutPromise({request:e,logs:t,handler:a}){let s;return{promise:new Promise(t=>{s=setTimeout(async()=>{t(await a.cacheMatch(e))},1e3*this._networkTimeoutSeconds)}),id:s}}async _getNetworkPromise({timeoutId:e,request:t,logs:a,handler:s}){let r,n;try{n=await s.fetchAndCachePut(t)}catch(e){e instanceof Error&&(r=e)}return e&&clearTimeout(e),(r||!n)&&(n=await s.cacheMatch(t)),n}},et=class extends Z{_networkTimeoutSeconds;constructor(e={}){super(e),this._networkTimeoutSeconds=e.networkTimeoutSeconds||0}async _handle(e,t){let a,s;try{let a=[t.fetch(e)];if(this._networkTimeoutSeconds){let e=h(1e3*this._networkTimeoutSeconds);a.push(e)}if(!(s=await Promise.race(a)))throw Error(`Timed out the network response after ${this._networkTimeoutSeconds} seconds.`)}catch(e){e instanceof Error&&(a=e)}if(!s)throw new l("no-response",{url:e.url,error:a});return s}};let ea=e=>e&&"object"==typeof e?e:{handle:e};var es=class{handler;match;method;catchHandler;constructor(e,t,a="GET"){this.handler=ea(t),this.match=e,this.method=a}setCatchHandler(e){this.catchHandler=ea(e)}},er=class e extends Z{_fallbackToNetwork;static defaultPrecacheCacheabilityPlugin={cacheWillUpdate:async({response:e})=>!e||e.status>=400?null:e};static copyRedirectedCacheableResponsesPlugin={cacheWillUpdate:async({response:e})=>e.redirected?await M(e):e};constructor(t={}){t.cacheName=c(t.cacheName),super(t),this._fallbackToNetwork=!1!==t.fallbackToNetwork,this.plugins.push(e.copyRedirectedCacheableResponsesPlugin)}async _handle(e,t){let a=await t.getPreloadResponse();if(a)return a;let s=await t.cacheMatch(e);return s||(t.event&&"install"===t.event.type?await this._handleInstall(e,t):await this._handleFetch(e,t))}async _handleFetch(e,t){let a,s=t.params||{};if(this._fallbackToNetwork){let r=s.integrity,n=e.integrity,i=!n||n===r;a=await t.fetch(new Request(e,{integrity:"no-cors"!==e.mode?n||r:void 0})),r&&i&&"no-cors"!==e.mode&&(this._useDefaultCacheabilityPluginIfNeeded(),await t.cachePut(e,a.clone()))}else throw new l("missing-precache-entry",{cacheName:this.cacheName,url:e.url});return a}async _handleInstall(e,t){this._useDefaultCacheabilityPluginIfNeeded();let a=await t.fetch(e);if(!await t.cachePut(e,a.clone()))throw new l("bad-precaching-response",{url:e.url,status:a.status});return a}_useDefaultCacheabilityPluginIfNeeded(){let t=null,a=0;for(let[s,r]of this.plugins.entries())r!==e.copyRedirectedCacheableResponsesPlugin&&(r===e.defaultPrecacheCacheabilityPlugin&&(t=s),r.cacheWillUpdate&&a++);0===a?this.plugins.push(e.defaultPrecacheCacheabilityPlugin):a>1&&null!==t&&this.plugins.splice(t,1)}},en=class extends es{_allowlist;_denylist;constructor(e,{allowlist:t=[/./],denylist:a=[]}={}){super(e=>this._match(e),e),this._allowlist=t,this._denylist=a}_match({url:e,request:t}){if(t&&"navigate"!==t.mode)return!1;let a=e.pathname+e.search;for(let e of this._denylist)if(e.test(a))return!1;return!!this._allowlist.some(e=>e.test(a))}},ei=class extends es{constructor(e,t,a){super(({url:t})=>{let a=e.exec(t.href);if(a)return t.origin!==location.origin&&0!==a.index?void 0:a.slice(1)},t,a)}};let ec=e=>{if(!e)throw new l("add-to-cache-list-unexpected-type",{entry:e});if("string"==typeof e){let t=new URL(e,location.href);return{cacheKey:t.href,url:t.href}}let{revision:t,url:a}=e;if(!a)throw new l("add-to-cache-list-unexpected-type",{entry:e});if(!t){let e=new URL(a,location.href);return{cacheKey:e.href,url:e.href}}let s=new URL(a,location.href),r=new URL(a,location.href);return s.searchParams.set("__WB_REVISION__",t),{cacheKey:s.href,url:r.href}};var eo=class{updatedURLs=[];notUpdatedURLs=[];handlerWillStart=async({request:e,state:t})=>{t&&(t.originalRequest=e)};cachedResponseWillBeUsed=async({event:e,state:t,cachedResponse:a})=>{if("install"===e.type&&t?.originalRequest&&t.originalRequest instanceof Request){let e=t.originalRequest.url;a?this.notUpdatedURLs.push(e):this.updatedURLs.push(e)}return a}};let el=async(e,t,a)=>{let s=t.map((e,t)=>({index:t,item:e})),r=async e=>{let t=[];for(;;){let r=s.pop();if(!r)return e(t);let n=await a(r.item);t.push({result:n,index:r.index})}},n=Array.from({length:e},()=>new Promise(r));return(await Promise.all(n)).flat().sort((e,t)=>e.index<t.index?-1:1).map(e=>e.result)};"u">typeof navigator&&/^((?!chrome|android).)*safari/i.test(navigator.userAgent);let eh="cache-entries",eu=e=>{let t=new URL(e,location.href);return t.hash="",t.href};var ed=class{_cacheName;_db=null;constructor(e){this._cacheName=e}_getId(e){return`${this._cacheName}|${eu(e)}`}_upgradeDb(e){let t=e.createObjectStore(eh,{keyPath:"id"});t.createIndex("cacheName","cacheName",{unique:!1}),t.createIndex("timestamp","timestamp",{unique:!1})}_upgradeDbAndDeleteOldDbs(e){this._upgradeDb(e),this._cacheName&&function(e,{blocked:t}={}){let a=indexedDB.deleteDatabase(e);t&&a.addEventListener("blocked",e=>t(e.oldVersion,e)),R(a).then(()=>void 0)}(this._cacheName)}async setTimestamp(e,t){e=eu(e);let a={id:this._getId(e),cacheName:this._cacheName,url:e,timestamp:t},s=(await this.getDb()).transaction(eh,"readwrite",{durability:"relaxed"});await s.store.put(a),await s.done}async getTimestamp(e){return(await (await this.getDb()).get(eh,this._getId(e)))?.timestamp}async expireEntries(e,t){let a=await (await this.getDb()).transaction(eh,"readwrite").store.index("timestamp").openCursor(null,"prev"),s=[],r=0;for(;a;){let n=a.value;n.cacheName===this._cacheName&&(e&&n.timestamp<e||t&&r>=t?(a.delete(),s.push(n.url)):r++),a=await a.continue()}return s}async getDb(){return this._db||(this._db=await S("serwist-expiration",1,{upgrade:this._upgradeDbAndDeleteOldDbs.bind(this)})),this._db}},em=class{_isRunning=!1;_rerunRequested=!1;_maxEntries;_maxAgeSeconds;_matchOptions;_cacheName;_timestampModel;constructor(e,t={}){this._maxEntries=t.maxEntries,this._maxAgeSeconds=t.maxAgeSeconds,this._matchOptions=t.matchOptions,this._cacheName=e,this._timestampModel=new ed(e)}async expireEntries(){if(this._isRunning){this._rerunRequested=!0;return}this._isRunning=!0;let e=this._maxAgeSeconds?Date.now()-1e3*this._maxAgeSeconds:0,t=await this._timestampModel.expireEntries(e,this._maxEntries),a=await self.caches.open(this._cacheName);for(let e of t)await a.delete(e,this._matchOptions);this._isRunning=!1,this._rerunRequested&&(this._rerunRequested=!1,this.expireEntries())}async updateTimestamp(e){await this._timestampModel.setTimestamp(e,Date.now())}async isURLExpired(e){if(!this._maxAgeSeconds)return!1;let t=await this._timestampModel.getTimestamp(e),a=Date.now()-1e3*this._maxAgeSeconds;return void 0===t||t<a}async delete(){this._rerunRequested=!1,await this._timestampModel.expireEntries(1/0)}},ef=class{_config;_cacheExpirations;constructor(e={}){this._config=e,this._cacheExpirations=new Map,this._config.maxAgeFrom||(this._config.maxAgeFrom="last-fetched"),this._config.purgeOnQuotaError&&(e=>{u.add(e)})(()=>this.deleteCacheAndMetadata())}_getCacheExpiration(e){if(e===o())throw new l("expire-custom-caches-only");let t=this._cacheExpirations.get(e);return t||(t=new em(e,this._config),this._cacheExpirations.set(e,t)),t}cachedResponseWillBeUsed({event:e,cacheName:t,request:a,cachedResponse:s}){if(!s)return null;let r=this._isResponseDateFresh(s),n=this._getCacheExpiration(t),i="last-used"===this._config.maxAgeFrom,c=(async()=>{i&&await n.updateTimestamp(a.url),await n.expireEntries()})();try{e.waitUntil(c)}catch{}return r?s:null}_isResponseDateFresh(e){if("last-used"===this._config.maxAgeFrom)return!0;let t=Date.now();if(!this._config.maxAgeSeconds)return!0;let a=this._getDateHeaderTimestamp(e);return null===a||a>=t-1e3*this._config.maxAgeSeconds}_getDateHeaderTimestamp(e){if(!e.headers.has("date"))return null;let t=new Date(e.headers.get("date")).getTime();return Number.isNaN(t)?null:t}async cacheDidUpdate({cacheName:e,request:t}){let a=this._getCacheExpiration(e);await a.updateTimestamp(t.url),await a.expireEntries()}async deleteCacheAndMetadata(){for(let[e,t]of this._cacheExpirations)await self.caches.delete(e),await t.delete();this._cacheExpirations=new Map}};let eg=/^\/(\w+\/)?collect/,ew=({serwist:e,cacheName:t,...a})=>{let s,r,c=t||i(n.googleAnalytics),o=new z("serwist-google-analytics",{maxRetentionTime:2880,onSync:async({queue:e})=>{let t;for(;t=await e.shiftRequest();){let{request:s,timestamp:r}=t,n=new URL(s.url);try{let e="POST"===s.method?new URLSearchParams(await s.clone().text()):n.searchParams,t=r-(Number(e.get("qt"))||0),i=Date.now()-t;if(e.set("qt",String(i)),a.parameterOverrides)for(let t of Object.keys(a.parameterOverrides)){let s=a.parameterOverrides[t];e.set(t,s)}"function"==typeof a.hitFilter&&a.hitFilter.call(null,e),await fetch(new Request(n.origin+n.pathname,{body:e.toString(),method:"POST",mode:"cors",credentials:"omit",headers:{"Content-Type":"text/plain"}}))}catch(a){throw await e.unshiftRequest(t),a}}}});for(let t of[new es(({url:e})=>"www.googletagmanager.com"===e.hostname&&"/gtm.js"===e.pathname,new ee({cacheName:c}),"GET"),new es(({url:e})=>"www.google-analytics.com"===e.hostname&&"/analytics.js"===e.pathname,new ee({cacheName:c}),"GET"),new es(({url:e})=>"www.googletagmanager.com"===e.hostname&&"/gtag/js"===e.pathname,new ee({cacheName:c}),"GET"),new es(s=({url:e})=>"www.google-analytics.com"===e.hostname&&eg.test(e.pathname),r=new et({plugins:[o]}),"GET"),new es(s,r,"POST")])e.registerRoute(t)};var ep=class{_fallbackUrls;_serwist;constructor({fallbackUrls:e,serwist:t}){this._fallbackUrls=e,this._serwist=t}async handlerDidError(e){for(let t of this._fallbackUrls)if("string"==typeof t){let e=await this._serwist.matchPrecache(t);if(void 0!==e)return e}else if(t.matcher(e)){let e=await this._serwist.matchPrecache(t.url);if(void 0!==e)return e}}};let ey=async(e,t)=>{try{if(206===t.status)return t;let a=e.headers.get("range");if(!a)throw new l("no-range-header");let s=(e=>{let t=e.trim().toLowerCase();if(!t.startsWith("bytes="))throw new l("unit-must-be-bytes",{normalizedRangeHeader:t});if(t.includes(","))throw new l("single-range-only",{normalizedRangeHeader:t});let a=/(\d*)-(\d*)/.exec(t);if(!a||!(a[1]||a[2]))throw new l("invalid-range-values",{normalizedRangeHeader:t});return{start:""===a[1]?void 0:Number(a[1]),end:""===a[2]?void 0:Number(a[2])}})(a),r=await t.blob(),n=((e,t,a)=>{let s,r,n=e.size;if(a&&a>n||t&&t<0)throw new l("range-not-satisfiable",{size:n,end:a,start:t});return void 0!==t&&void 0!==a?(s=t,r=a+1):void 0!==t&&void 0===a?(s=t,r=n):void 0!==a&&void 0===t&&(s=n-a,r=n),{start:s,end:r}})(r,s.start,s.end),i=r.slice(n.start,n.end),c=i.size,o=new Response(i,{status:206,statusText:"Partial Content",headers:t.headers});return o.headers.set("Content-Length",String(c)),o.headers.set("Content-Range",`bytes ${n.start}-${n.end-1}/${r.size}`),o}catch(e){return new Response("",{status:416,statusText:"Range Not Satisfiable"})}};var e_=class{cachedResponseWillBeUsed=async({request:e,cachedResponse:t})=>t&&e.headers.has("range")?await ey(e,t):t},eb=class extends Z{async _handle(e,t){let a,s=await t.cacheMatch(e);if(s);else try{s=await t.fetchAndCachePut(e)}catch(e){e instanceof Error&&(a=e)}if(!s)throw new l("no-response",{url:e.url,error:a});return s}},ex=class extends Z{constructor(e={}){super(e),this.plugins.some(e=>"cacheWillUpdate"in e)||this.plugins.unshift(J)}async _handle(e,t){let a,s=t.fetchAndCachePut(e).catch(()=>{});t.waitUntil(s);let r=await t.cacheMatch(e);if(r);else try{r=await s}catch(e){e instanceof Error&&(a=e)}if(!r)throw new l("no-response",{url:e.url,error:a});return r}},ev=class extends es{constructor(e,t){super(({request:a})=>{let s=e.getUrlsToPrecacheKeys();for(let r of function*(e,{directoryIndex:t="index.html",ignoreURLParametersMatching:a=[/^utm_/,/^fbclid$/],cleanURLs:s=!0,urlManipulation:r}={}){let n=new URL(e,location.href);n.hash="",yield n.href;let i=((e,t=[])=>{for(let a of[...e.searchParams.keys()])t.some(e=>e.test(a))&&e.searchParams.delete(a);return e})(n,a);if(yield i.href,t&&i.pathname.endsWith("/")){let e=new URL(i.href);e.pathname+=t,yield e.href}if(s){let e=new URL(i.href);e.pathname+=".html",yield e.href}if(r)for(let e of r({url:n}))yield e.href}(a.url,t)){let t=s.get(r);if(t)return{cacheKey:t,integrity:e.getIntegrityForPrecacheKey(t)}}},e.precacheStrategy)}},eE=class{_precacheController;constructor({precacheController:e}){this._precacheController=e}cacheKeyWillBeUsed=async({request:e,params:t})=>{let a=t?.cacheKey||this._precacheController.getPrecacheKeyForUrl(e.url);return a?new Request(a,{headers:e.headers}):e}},eR=class{_urlsToCacheKeys=new Map;_urlsToCacheModes=new Map;_cacheKeysToIntegrities=new Map;_concurrentPrecaching;_precacheStrategy;_routes;_defaultHandlerMap;_catchHandler;_requestRules;constructor({precacheEntries:e,precacheOptions:t,skipWaiting:a=!1,importScripts:s,navigationPreload:r=!1,cacheId:i,clientsClaim:o=!1,runtimeCaching:l,offlineAnalyticsConfig:h,disableDevLogs:u=!1,fallbacks:d,requestRules:m}={}){const{precacheStrategyOptions:f,precacheRouteOptions:g,precacheMiscOptions:w}=((e,t={})=>{let{cacheName:a,plugins:s=[],fetchOptions:r,matchOptions:n,fallbackToNetwork:i,directoryIndex:o,ignoreURLParametersMatching:l,cleanURLs:h,urlManipulation:u,cleanupOutdatedCaches:d,concurrency:m=10,navigateFallback:f,navigateFallbackAllowlist:g,navigateFallbackDenylist:w}=t??{};return{precacheStrategyOptions:{cacheName:c(a),plugins:[...s,new eE({precacheController:e})],fetchOptions:r,matchOptions:n,fallbackToNetwork:i},precacheRouteOptions:{directoryIndex:o,ignoreURLParametersMatching:l,cleanURLs:h,urlManipulation:u},precacheMiscOptions:{cleanupOutdatedCaches:d,concurrency:m,navigateFallback:f,navigateFallbackAllowlist:g,navigateFallbackDenylist:w}}})(this,t);if(this._concurrentPrecaching=w.concurrency,this._precacheStrategy=new er(f),this._routes=new Map,this._defaultHandlerMap=new Map,this._requestRules=m,this.handleInstall=this.handleInstall.bind(this),this.handleActivate=this.handleActivate.bind(this),this.handleFetch=this.handleFetch.bind(this),this.handleCache=this.handleCache.bind(this),s&&s.length>0&&self.importScripts(...s),r&&self.registration?.navigationPreload&&self.addEventListener("activate",e=>{e.waitUntil(self.registration.navigationPreload.enable().then(()=>{}))}),void 0!==i&&(e=>{var t=e;for(let e of Object.keys(n))(e=>{let a=t[e];"string"==typeof a&&(n[e]=a)})(e)})({prefix:i}),a?self.skipWaiting():self.addEventListener("message",e=>{e.data&&"SKIP_WAITING"===e.data.type&&self.skipWaiting()}),o&&self.addEventListener("activate",()=>self.clients.claim()),e&&e.length>0&&this.addToPrecacheList(e),w.cleanupOutdatedCaches&&(e=>{self.addEventListener("activate",t=>{t.waitUntil(p(c(e)).then(e=>{}))})})(f.cacheName),this.registerRoute(new ev(this,g)),w.navigateFallback&&this.registerRoute(new en(this.createHandlerBoundToUrl(w.navigateFallback),{allowlist:w.navigateFallbackAllowlist,denylist:w.navigateFallbackDenylist})),void 0!==h&&("boolean"==typeof h?h&&ew({serwist:this}):ew({...h,serwist:this})),void 0!==l){if(void 0!==d){const e=new ep({fallbackUrls:d.entries,serwist:this});l.forEach(t=>{t.handler instanceof Z&&!t.handler.plugins.some(e=>"handlerDidError"in e)&&t.handler.plugins.push(e)})}for(const e of l)this.registerCapture(e.matcher,e.handler,e.method)}u&&(self.__WB_DISABLE_DEV_LOGS=!0)}get precacheStrategy(){return this._precacheStrategy}get routes(){return this._routes}addEventListeners(){self.addEventListener("install",this.handleInstall),self.addEventListener("activate",this.handleActivate),self.addEventListener("fetch",this.handleFetch),self.addEventListener("message",this.handleCache)}addToPrecacheList(e){let t=[];for(let a of e){"string"==typeof a?t.push(a):a&&!a.integrity&&void 0===a.revision&&t.push(a.url);let{cacheKey:e,url:s}=ec(a),r="string"!=typeof a&&a.revision?"reload":"default";if(this._urlsToCacheKeys.has(s)&&this._urlsToCacheKeys.get(s)!==e)throw new l("add-to-cache-list-conflicting-entries",{firstEntry:this._urlsToCacheKeys.get(s),secondEntry:e});if("string"!=typeof a&&a.integrity){if(this._cacheKeysToIntegrities.has(e)&&this._cacheKeysToIntegrities.get(e)!==a.integrity)throw new l("add-to-cache-list-conflicting-integrities",{url:s});this._cacheKeysToIntegrities.set(e,a.integrity)}this._urlsToCacheKeys.set(s,e),this._urlsToCacheModes.set(s,r)}t.length>0&&console.warn(`Serwist is precaching URLs without revision info: ${t.join(", ")}
-This is generally NOT safe. Learn more at https://bit.ly/wb-precache`)}handleInstall(e){return this.registerRequestRules(e),y(e,async()=>{let t=new eo;this.precacheStrategy.plugins.push(t),await el(this._concurrentPrecaching,Array.from(this._urlsToCacheKeys.entries()),async([t,a])=>{let s=this._cacheKeysToIntegrities.get(a),r=this._urlsToCacheModes.get(t),n=new Request(t,{integrity:s,cache:r,credentials:"same-origin"});await Promise.all(this.precacheStrategy.handleAll({event:e,request:n,url:new URL(n.url),params:{cacheKey:a}}))});let{updatedURLs:a,notUpdatedURLs:s}=t;return{updatedURLs:a,notUpdatedURLs:s}})}async registerRequestRules(e){if(this._requestRules&&e?.addRoutes)try{await e.addRoutes(this._requestRules),this._requestRules=void 0}catch(e){throw e}}handleActivate(e){return y(e,async()=>{let e=await self.caches.open(this.precacheStrategy.cacheName),t=await e.keys(),a=new Set(this._urlsToCacheKeys.values()),s=[];for(let r of t)a.has(r.url)||(await e.delete(r),s.push(r.url));return{deletedCacheRequests:s}})}handleFetch(e){let{request:t}=e,a=this.handleRequest({request:t,event:e});a&&e.respondWith(a)}handleCache(e){if(e.data&&"CACHE_URLS"===e.data.type){let{payload:t}=e.data,a=Promise.all(t.urlsToCache.map(t=>{let a;return a="string"==typeof t?new Request(t):new Request(...t),this.handleRequest({request:a,event:e})}));e.waitUntil(a),e.ports?.[0]&&a.then(()=>e.ports[0].postMessage(!0))}}setDefaultHandler(e,t="GET"){this._defaultHandlerMap.set(t,ea(e))}setCatchHandler(e){this._catchHandler=ea(e)}registerCapture(e,t,a){let s=((e,t,a)=>{if("string"==typeof e){let s=new URL(e,location.href);return new es(({url:e})=>e.href===s.href,t,a)}if(e instanceof RegExp)return new ei(e,t,a);if("function"==typeof e)return new es(e,t,a);if(e instanceof es)return e;throw new l("unsupported-route-type",{moduleName:"serwist",funcName:"parseRoute",paramName:"capture"})})(e,t,a);return this.registerRoute(s),s}registerRoute(e){this._routes.has(e.method)||this._routes.set(e.method,[]),this._routes.get(e.method).push(e)}unregisterRoute(e){if(!this._routes.has(e.method))throw new l("unregister-route-but-not-found-with-method",{method:e.method});let t=this._routes.get(e.method).indexOf(e);if(t>-1)this._routes.get(e.method).splice(t,1);else throw new l("unregister-route-route-not-registered")}getUrlsToPrecacheKeys(){return this._urlsToCacheKeys}getPrecachedUrls(){return[...this._urlsToCacheKeys.keys()]}getPrecacheKeyForUrl(e){let t=new URL(e,location.href);return this._urlsToCacheKeys.get(t.href)}getIntegrityForPrecacheKey(e){return this._cacheKeysToIntegrities.get(e)}async matchPrecache(e){let t=e instanceof Request?e.url:e,a=this.getPrecacheKeyForUrl(t);if(a)return(await self.caches.open(this.precacheStrategy.cacheName)).match(a)}createHandlerBoundToUrl(e){let t=this.getPrecacheKeyForUrl(e);if(!t)throw new l("non-precached-url",{url:e});return a=>(a.request=new Request(e),a.params={cacheKey:t,...a.params},this.precacheStrategy.handle(a))}handleRequest({request:e,event:t}){let a,s=new URL(e.url,location.href);if(!s.protocol.startsWith("http"))return;let r=s.origin===location.origin,{params:n,route:i}=this.findMatchingRoute({event:t,request:e,sameOrigin:r,url:s}),c=i?.handler,o=e.method;if(!c&&this._defaultHandlerMap.has(o)&&(c=this._defaultHandlerMap.get(o)),!c)return;try{a=c.handle({url:s,request:e,event:t,params:n})}catch(e){a=Promise.reject(e)}let l=i?.catchHandler;return a instanceof Promise&&(this._catchHandler||l)&&(a=a.catch(async a=>{if(l)try{return await l.handle({url:s,request:e,event:t,params:n})}catch(e){e instanceof Error&&(a=e)}if(this._catchHandler)return this._catchHandler.handle({url:s,request:e,event:t});throw a})),a}findMatchingRoute({url:e,sameOrigin:t,request:a,event:s}){for(let r of this._routes.get(a.method)||[]){let n,i=r.match({url:e,sameOrigin:t,request:a,event:s});if(i)return Array.isArray(n=i)&&0===n.length||i.constructor===Object&&0===Object.keys(i).length?n=void 0:"boolean"==typeof i&&(n=void 0),{route:r,params:n}}return{}}};let eq=[{matcher:/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,handler:new eb({cacheName:"google-fonts-webfonts",plugins:[new ef({maxEntries:4,maxAgeSeconds:31536e3,maxAgeFrom:"last-used"})]})},{matcher:/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,handler:new ex({cacheName:"google-fonts-stylesheets",plugins:[new ef({maxEntries:4,maxAgeSeconds:604800,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,handler:new ex({cacheName:"static-font-assets",plugins:[new ef({maxEntries:4,maxAgeSeconds:604800,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,handler:new ex({cacheName:"static-image-assets",plugins:[new ef({maxEntries:64,maxAgeSeconds:2592e3,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/static.+\.js$/i,handler:new eb({cacheName:"next-static-js-assets",plugins:[new ef({maxEntries:64,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/image\?url=.+$/i,handler:new ex({cacheName:"next-image",plugins:[new ef({maxEntries:64,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:mp3|wav|ogg)$/i,handler:new eb({cacheName:"static-audio-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"}),new e_]})},{matcher:/\.(?:mp4|webm)$/i,handler:new eb({cacheName:"static-video-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"}),new e_]})},{matcher:/\.(?:js)$/i,handler:new ex({cacheName:"static-js-assets",plugins:[new ef({maxEntries:48,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:css|less)$/i,handler:new ex({cacheName:"static-style-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/data\/.+\/.+\.json$/i,handler:new ee({cacheName:"next-data",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:json|xml|csv)$/i,handler:new ee({cacheName:"static-data-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/api\/auth\/.*/,handler:new et({networkTimeoutSeconds:10})},{matcher:({sameOrigin:e,url:{pathname:t}})=>e&&t.startsWith("/api/"),method:"GET",handler:new ee({cacheName:"apis",plugins:[new ef({maxEntries:16,maxAgeSeconds:86400,maxAgeFrom:"last-used"})],networkTimeoutSeconds:10})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages-rsc-prefetch",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>"1"===e.headers.get("RSC")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages-rsc",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>e.headers.get("Content-Type")?.includes("text/html")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({url:{pathname:e},sameOrigin:t})=>t&&!e.startsWith("/api/"),handler:new ee({cacheName:"others",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({sameOrigin:e})=>!e,handler:new ee({cacheName:"cross-origin",plugins:[new ef({maxEntries:32,maxAgeSeconds:3600})],networkTimeoutSeconds:10})},{matcher:/.*/i,method:"GET",handler:new et}];async function eS(e){let t,a=(t=globalThis.navigator)&&"function"==typeof t.setAppBadge?t:null;if(a)try{e>0?await a.setAppBadge?.(e):await a.clearAppBadge?.()}catch{}}let eD="app-badge",eN="/__badge_count";async function eC(){try{let e=await caches.open(eD),t=await e.match(eN);if(!t)return 0;let a=Number(await t.text());return Number.isFinite(a)&&a>0?a:0}catch{return 0}}async function eT(e){try{let t=await caches.open(eD);await t.put(eN,new Response(String(Math.max(0,e))))}catch{}}let eP=new eR({precacheEntries:[{'revision':'9771bcf20ba38f90712d273569b03e75','url':'/_next/static/a_cJF6nYmUf2MP8Dc_Uru/_buildManifest.js'},{'revision':'b6652df95db52feb4daf4eca35380933','url':'/_next/static/a_cJF6nYmUf2MP8Dc_Uru/_ssgManifest.js'},{'revision':null,'url':'/_next/static/chunks/239-a80c4269a471e906.js'},{'revision':null,'url':'/_next/static/chunks/273-286e8118997487e4.js'},{'revision':null,'url':'/_next/static/chunks/502-51615e654ca3e238.js'},{'revision':null,'url':'/_next/static/chunks/64-52fe3e9f730a5866.js'},{'revision':null,'url':'/_next/static/chunks/698-dfab1f631472744b.js'},{'revision':null,'url':'/_next/static/chunks/833-ddfd79463d14cc11.js'},{'revision':null,'url':'/_next/static/chunks/87c73c54-f46fec743414da25.js'},{'revision':null,'url':'/_next/static/chunks/960-bffd61e35e8449db.js'},{'revision':null,'url':'/_next/static/chunks/app/_global-error/page-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/app/_not-found/page-1f43d9cfe782d782.js'},{'revision':null,'url':'/_next/static/chunks/app/auth/success/page-070ae95134c3f3e7.js'},{'revision':null,'url':'/_next/static/chunks/app/debug/page-8b906d41cb385b1c.js'},{'revision':null,'url':'/_next/static/chunks/app/install/page-4a619ab781fc3914.js'},{'revision':null,'url':'/_next/static/chunks/app/join/layout-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/app/join/page-ffd179b2863f3f66.js'},{'revision':null,'url':'/_next/static/chunks/app/layout-32c2bc6ee5f243ac.js'},{'revision':null,'url':'/_next/static/chunks/app/link/page-e1a00d7571240d6f.js'},{'revision':null,'url':'/_next/static/chunks/app/logout/page-1474173a72f77bf9.js'},{'revision':null,'url':'/_next/static/chunks/app/me/page-08898f94ae9de759.js'},{'revision':null,'url':'/_next/static/chunks/app/notifications/page-7ca63df0f65a3931.js'},{'revision':null,'url':'/_next/static/chunks/app/orders/page-bdc41bbeeb9b5d6a.js'},{'revision':null,'url':'/_next/static/chunks/app/overview/page-5147087d24205b8b.js'},{'revision':null,'url':'/_next/static/chunks/app/page-64056cb57b7df288.js'},{'revision':null,'url':'/_next/static/chunks/app/register/page-4403385d51642139.js'},{'revision':null,'url':'/_next/static/chunks/app/settlements/page-fe9be6e718a402b6.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/c/%5Bid%5D/page-c03446d06f66d1f7.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/flash/page-c5d4e48b33222252.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/food-train/page-a0f77ec528b8b941.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/page-efc73e925dbb5979.js'},{'revision':null,'url':'/_next/static/chunks/app/spot/%5Bid%5D/page-af56c94bf6cf4248.js'},{'revision':null,'url':'/_next/static/chunks/app/spot/page-8357df49c76fe5c7.js'},{'revision':null,'url':'/_next/static/chunks/app/wallet/page-1655593dcfe0469d.js'},{'revision':null,'url':'/_next/static/chunks/framework-369218c9910949e7.js'},{'revision':null,'url':'/_next/static/chunks/main-app-b7c599722f79bdc1.js'},{'revision':null,'url':'/_next/static/chunks/main-be79bb95d5b42faf.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/app-error-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/forbidden-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/global-error-ee767baf43a5a2ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/not-found-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/unauthorized-27860661a41da3ba.js'},{'revision':'846118c33b2c0e922d7b3a7676f81f6f','url':'/_next/static/chunks/polyfills-42372ed130431b0a.js'},{'revision':null,'url':'/_next/static/chunks/webpack-b695f13b2836b942.js'},{'revision':null,'url':'/_next/static/css/69da13f363dbf0f9.css'},{'revision':'3fc49c2a6022fd7a30dd0fd643f57866','url':'/brand/banner.jpg'},{'revision':'27ce8d11af1fa5310a38e0a873b612c5','url':'/brand/logo.jpg'},{'revision':'21c8674ef5de87f15149b554bef954e8','url':'/icons/android/launchericon-144x144.png'},{'revision':'5a9f92af94660d63c0c436694774f624','url':'/icons/android/launchericon-192x192.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/android/launchericon-48x48.png'},{'revision':'5cc74a1180e9da94973b047042aded54','url':'/icons/android/launchericon-512x512.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/android/launchericon-72x72.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/android/launchericon-96x96.png'},{'revision':'7ee5446d081467e5800848c4ce3da6f3','url':'/icons/ios/100.png'},{'revision':'f11d312cd3d7cfee6f02a23906fd108f','url':'/icons/ios/1024.png'},{'revision':'9816127b8bdc3776952711b0977b2ef8','url':'/icons/ios/114.png'},{'revision':'0d05e20f509d2d38b37f0a3c101aa35d','url':'/icons/ios/120.png'},{'revision':'b98a05d66f7494066d0ba3a42aeca1ae','url':'/icons/ios/128.png'},{'revision':'2e0a67e030ee7061d57eb221fe780723','url':'/icons/ios/144.png'},{'revision':'e869e4705bd09e9b83a01afb1bd376d9','url':'/icons/ios/152.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/ios/16.png'},{'revision':'ea8d9c9d88c17d23c42850625c442fa0','url':'/icons/ios/167.png'},{'revision':'b6707a5ebfd9f53b3c6a9dc90a4082e0','url':'/icons/ios/180.png'},{'revision':'7fdcee2509439b092e668c9448cf9053','url':'/icons/ios/192.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/ios/20.png'},{'revision':'6d066b4ef4ee5b5dba9621aab8cc9a1a','url':'/icons/ios/256.png'},{'revision':'9005c4765e0f8e17d2eb63e028c24bf1','url':'/icons/ios/29.png'},{'revision':'19391d3de8599244346b47ad599d83e5','url':'/icons/ios/32.png'},{'revision':'0ce2683b3ee1866903723d275c6349fa','url':'/icons/ios/40.png'},{'revision':'629c5674e8fd89d33f57aaf092894285','url':'/icons/ios/50.png'},{'revision':'8cd35511a222d5cc4e4a23ef4c07688b','url':'/icons/ios/512.png'},{'revision':'34a8adcc3ef02c6cc74f86449481ac4d','url':'/icons/ios/57.png'},{'revision':'f9a363efb255f09f867333414e13458f','url':'/icons/ios/58.png'},{'revision':'eca69c492f9525f6661f80e844f89927','url':'/icons/ios/60.png'},{'revision':'324b63561172d305fc2f7021f71a527c','url':'/icons/ios/64.png'},{'revision':'2160847c205affc9d6c2850a9e67fdec','url':'/icons/ios/72.png'},{'revision':'b5c64f7e180c5b51e41deded738ed5ae','url':'/icons/ios/76.png'},{'revision':'254abe424ab5f8498480524eae0de96d','url':'/icons/ios/80.png'},{'revision':'e7fa488c29c40d108ec85e1f82a79d31','url':'/icons/ios/87.png'},{'revision':'84c7cad94d49288c8e93ba7e2679b33d','url':'/icons/windows/LargeTile.scale-100.png'},{'revision':'432a84dd02c63646e859edbb40df190c','url':'/icons/windows/LargeTile.scale-125.png'},{'revision':'096081a71cb71daa44c05e050695d8de','url':'/icons/windows/LargeTile.scale-150.png'},{'revision':'7608704a00592217d63f7b318c2e5f38','url':'/icons/windows/LargeTile.scale-200.png'},{'revision':'6e17ba59386ded6be1f10da2e5bfb9e0','url':'/icons/windows/LargeTile.scale-400.png'},{'revision':'3caa90776f06fc3ac443417ff6f9e9c2','url':'/icons/windows/SmallTile.scale-100.png'},{'revision':'267cadd8270e32c155b6dd3cd2af81b3','url':'/icons/windows/SmallTile.scale-125.png'},{'revision':'11c0fd9522841246d62af0fd09530459','url':'/icons/windows/SmallTile.scale-150.png'},{'revision':'d9596560b5df161ad68603a063323dea','url':'/icons/windows/SmallTile.scale-200.png'},{'revision':'fc07ec25a8ba517f6d69d36fc2bcfb5c','url':'/icons/windows/SmallTile.scale-400.png'},{'revision':'027040ff0058a27ef9bb9d957f0e1112','url':'/icons/windows/SplashScreen.scale-100.png'},{'revision':'1fc60c21b8e911b576605222e4604cac','url':'/icons/windows/SplashScreen.scale-125.png'},{'revision':'d7835db7d741404674b7e04de3648f24','url':'/icons/windows/SplashScreen.scale-150.png'},{'revision':'4069bf20884b973f3cb2256297917aae','url':'/icons/windows/SplashScreen.scale-200.png'},{'revision':'ac93274617295e12e608752ba70c9d5e','url':'/icons/windows/SplashScreen.scale-400.png'},{'revision':'4abc934b1324b56e1f66abbe6f2b2212','url':'/icons/windows/Square150x150Logo.scale-100.png'},{'revision':'121cd7794166b37d0b05f475eb55cc65','url':'/icons/windows/Square150x150Logo.scale-125.png'},{'revision':'39d68da5beb84c82745b7e562bbc6113','url':'/icons/windows/Square150x150Logo.scale-150.png'},{'revision':'31e6b17966d5c6da8ff8c6f0a78ea945','url':'/icons/windows/Square150x150Logo.scale-200.png'},{'revision':'4d68017168a3e5bb2173af09fcea4a6d','url':'/icons/windows/Square150x150Logo.scale-400.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-96.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-96.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.scale-100.png'},{'revision':'898a72b8a205ca8683f07e44931c5b3f','url':'/icons/windows/Square44x44Logo.scale-125.png'},{'revision':'c25a1b65e3a12b62d154556826ed1d16','url':'/icons/windows/Square44x44Logo.scale-150.png'},{'revision':'267cadd8270e32c155b6dd3cd2af81b3','url':'/icons/windows/Square44x44Logo.scale-200.png'},{'revision':'5044ecfe554402a9e9ef514a4ce20b3f','url':'/icons/windows/Square44x44Logo.scale-400.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.targetsize-96.png'},{'revision':'06091ad8946496f71160e08f3acd694e','url':'/icons/windows/StoreLogo.scale-100.png'},{'revision':'47e34490c42fe1dae111b67301bf4b35','url':'/icons/windows/StoreLogo.scale-125.png'},{'revision':'bdd2d8c69c0be099276c4d6567fa5d9e','url':'/icons/windows/StoreLogo.scale-150.png'},{'revision':'111893dec64c28034160243677a0ffce','url':'/icons/windows/StoreLogo.scale-200.png'},{'revision':'9c66299aa54ba377775dc59fec049bde','url':'/icons/windows/StoreLogo.scale-400.png'},{'revision':'bc73480baeea24805571faccca15c354','url':'/icons/windows/Wide310x150Logo.scale-100.png'},{'revision':'77e3655ad0441ad2dd39395b2e641119','url':'/icons/windows/Wide310x150Logo.scale-125.png'},{'revision':'90b6c6c60ffa9081541da24cba21c3b3','url':'/icons/windows/Wide310x150Logo.scale-150.png'},{'revision':'027040ff0058a27ef9bb9d957f0e1112','url':'/icons/windows/Wide310x150Logo.scale-200.png'},{'revision':'4069bf20884b973f3cb2256297917aae','url':'/icons/windows/Wide310x150Logo.scale-400.png'},{'revision':'cc01aff3744d47049dea63726110b6e2','url':'/manifest.json'}],skipWaiting:!0,clientsClaim:!0,navigationPreload:!0,runtimeCaching:eq});self.addEventListener("push",e=>{let t=e.data?.json();if(!t)return;let a=t.title||"新訊息",s={body:t.body||"您有一則新通知",icon:"/icons/android/launchericon-192x192.png",badge:"/icons/android/launchericon-192x192.png",data:t.url||"/"};e.waitUntil((async()=>{await self.registration.showNotification(a,s);let e=await eC()+1;await eT(e),await eS(e)})())}),self.addEventListener("notificationclick",e=>{e.notification.close(),e.waitUntil(self.clients.openWindow(e.notification.data))}),eP.addEventListeners()})();
+This is generally NOT safe. Learn more at https://bit.ly/wb-precache`)}handleInstall(e){return this.registerRequestRules(e),y(e,async()=>{let t=new eo;this.precacheStrategy.plugins.push(t),await el(this._concurrentPrecaching,Array.from(this._urlsToCacheKeys.entries()),async([t,a])=>{let s=this._cacheKeysToIntegrities.get(a),r=this._urlsToCacheModes.get(t),n=new Request(t,{integrity:s,cache:r,credentials:"same-origin"});await Promise.all(this.precacheStrategy.handleAll({event:e,request:n,url:new URL(n.url),params:{cacheKey:a}}))});let{updatedURLs:a,notUpdatedURLs:s}=t;return{updatedURLs:a,notUpdatedURLs:s}})}async registerRequestRules(e){if(this._requestRules&&e?.addRoutes)try{await e.addRoutes(this._requestRules),this._requestRules=void 0}catch(e){throw e}}handleActivate(e){return y(e,async()=>{let e=await self.caches.open(this.precacheStrategy.cacheName),t=await e.keys(),a=new Set(this._urlsToCacheKeys.values()),s=[];for(let r of t)a.has(r.url)||(await e.delete(r),s.push(r.url));return{deletedCacheRequests:s}})}handleFetch(e){let{request:t}=e,a=this.handleRequest({request:t,event:e});a&&e.respondWith(a)}handleCache(e){if(e.data&&"CACHE_URLS"===e.data.type){let{payload:t}=e.data,a=Promise.all(t.urlsToCache.map(t=>{let a;return a="string"==typeof t?new Request(t):new Request(...t),this.handleRequest({request:a,event:e})}));e.waitUntil(a),e.ports?.[0]&&a.then(()=>e.ports[0].postMessage(!0))}}setDefaultHandler(e,t="GET"){this._defaultHandlerMap.set(t,ea(e))}setCatchHandler(e){this._catchHandler=ea(e)}registerCapture(e,t,a){let s=((e,t,a)=>{if("string"==typeof e){let s=new URL(e,location.href);return new es(({url:e})=>e.href===s.href,t,a)}if(e instanceof RegExp)return new ei(e,t,a);if("function"==typeof e)return new es(e,t,a);if(e instanceof es)return e;throw new l("unsupported-route-type",{moduleName:"serwist",funcName:"parseRoute",paramName:"capture"})})(e,t,a);return this.registerRoute(s),s}registerRoute(e){this._routes.has(e.method)||this._routes.set(e.method,[]),this._routes.get(e.method).push(e)}unregisterRoute(e){if(!this._routes.has(e.method))throw new l("unregister-route-but-not-found-with-method",{method:e.method});let t=this._routes.get(e.method).indexOf(e);if(t>-1)this._routes.get(e.method).splice(t,1);else throw new l("unregister-route-route-not-registered")}getUrlsToPrecacheKeys(){return this._urlsToCacheKeys}getPrecachedUrls(){return[...this._urlsToCacheKeys.keys()]}getPrecacheKeyForUrl(e){let t=new URL(e,location.href);return this._urlsToCacheKeys.get(t.href)}getIntegrityForPrecacheKey(e){return this._cacheKeysToIntegrities.get(e)}async matchPrecache(e){let t=e instanceof Request?e.url:e,a=this.getPrecacheKeyForUrl(t);if(a)return(await self.caches.open(this.precacheStrategy.cacheName)).match(a)}createHandlerBoundToUrl(e){let t=this.getPrecacheKeyForUrl(e);if(!t)throw new l("non-precached-url",{url:e});return a=>(a.request=new Request(e),a.params={cacheKey:t,...a.params},this.precacheStrategy.handle(a))}handleRequest({request:e,event:t}){let a,s=new URL(e.url,location.href);if(!s.protocol.startsWith("http"))return;let r=s.origin===location.origin,{params:n,route:i}=this.findMatchingRoute({event:t,request:e,sameOrigin:r,url:s}),c=i?.handler,o=e.method;if(!c&&this._defaultHandlerMap.has(o)&&(c=this._defaultHandlerMap.get(o)),!c)return;try{a=c.handle({url:s,request:e,event:t,params:n})}catch(e){a=Promise.reject(e)}let l=i?.catchHandler;return a instanceof Promise&&(this._catchHandler||l)&&(a=a.catch(async a=>{if(l)try{return await l.handle({url:s,request:e,event:t,params:n})}catch(e){e instanceof Error&&(a=e)}if(this._catchHandler)return this._catchHandler.handle({url:s,request:e,event:t});throw a})),a}findMatchingRoute({url:e,sameOrigin:t,request:a,event:s}){for(let r of this._routes.get(a.method)||[]){let n,i=r.match({url:e,sameOrigin:t,request:a,event:s});if(i)return Array.isArray(n=i)&&0===n.length||i.constructor===Object&&0===Object.keys(i).length?n=void 0:"boolean"==typeof i&&(n=void 0),{route:r,params:n}}return{}}};let eq=[{matcher:/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,handler:new eb({cacheName:"google-fonts-webfonts",plugins:[new ef({maxEntries:4,maxAgeSeconds:31536e3,maxAgeFrom:"last-used"})]})},{matcher:/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,handler:new ex({cacheName:"google-fonts-stylesheets",plugins:[new ef({maxEntries:4,maxAgeSeconds:604800,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,handler:new ex({cacheName:"static-font-assets",plugins:[new ef({maxEntries:4,maxAgeSeconds:604800,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,handler:new ex({cacheName:"static-image-assets",plugins:[new ef({maxEntries:64,maxAgeSeconds:2592e3,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/static.+\.js$/i,handler:new eb({cacheName:"next-static-js-assets",plugins:[new ef({maxEntries:64,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/image\?url=.+$/i,handler:new ex({cacheName:"next-image",plugins:[new ef({maxEntries:64,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:mp3|wav|ogg)$/i,handler:new eb({cacheName:"static-audio-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"}),new e_]})},{matcher:/\.(?:mp4|webm)$/i,handler:new eb({cacheName:"static-video-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"}),new e_]})},{matcher:/\.(?:js)$/i,handler:new ex({cacheName:"static-js-assets",plugins:[new ef({maxEntries:48,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:css|less)$/i,handler:new ex({cacheName:"static-style-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/_next\/data\/.+\/.+\.json$/i,handler:new ee({cacheName:"next-data",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\.(?:json|xml|csv)$/i,handler:new ee({cacheName:"static-data-assets",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400,maxAgeFrom:"last-used"})]})},{matcher:/\/api\/auth\/.*/,handler:new et({networkTimeoutSeconds:10})},{matcher:({sameOrigin:e,url:{pathname:t}})=>e&&t.startsWith("/api/"),method:"GET",handler:new ee({cacheName:"apis",plugins:[new ef({maxEntries:16,maxAgeSeconds:86400,maxAgeFrom:"last-used"})],networkTimeoutSeconds:10})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>"1"===e.headers.get("RSC")&&"1"===e.headers.get("Next-Router-Prefetch")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages-rsc-prefetch",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>"1"===e.headers.get("RSC")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages-rsc",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({request:e,url:{pathname:t},sameOrigin:a})=>e.headers.get("Content-Type")?.includes("text/html")&&a&&!t.startsWith("/api/"),handler:new ee({cacheName:"pages",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({url:{pathname:e},sameOrigin:t})=>t&&!e.startsWith("/api/"),handler:new ee({cacheName:"others",plugins:[new ef({maxEntries:32,maxAgeSeconds:86400})]})},{matcher:({sameOrigin:e})=>!e,handler:new ee({cacheName:"cross-origin",plugins:[new ef({maxEntries:32,maxAgeSeconds:3600})],networkTimeoutSeconds:10})},{matcher:/.*/i,method:"GET",handler:new et}];async function eS(e){let t,a=(t=globalThis.navigator)&&"function"==typeof t.setAppBadge?t:null;if(a)try{e>0?await a.setAppBadge?.(e):await a.clearAppBadge?.()}catch{}}let eD="app-badge",eN="/__badge_count";async function eC(){try{let e=await caches.open(eD),t=await e.match(eN);if(!t)return 0;let a=Number(await t.text());return Number.isFinite(a)&&a>0?a:0}catch{return 0}}async function eT(e){try{let t=await caches.open(eD);await t.put(eN,new Response(String(Math.max(0,e))))}catch{}}let eP=new eR({precacheEntries:[{'revision':null,'url':'/_next/static/chunks/239-a80c4269a471e906.js'},{'revision':null,'url':'/_next/static/chunks/273-286e8118997487e4.js'},{'revision':null,'url':'/_next/static/chunks/502-51615e654ca3e238.js'},{'revision':null,'url':'/_next/static/chunks/64-52fe3e9f730a5866.js'},{'revision':null,'url':'/_next/static/chunks/698-7f322e2e9bb5b7d2.js'},{'revision':null,'url':'/_next/static/chunks/833-ddfd79463d14cc11.js'},{'revision':null,'url':'/_next/static/chunks/87c73c54-f46fec743414da25.js'},{'revision':null,'url':'/_next/static/chunks/960-bffd61e35e8449db.js'},{'revision':null,'url':'/_next/static/chunks/app/_global-error/page-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/app/_not-found/page-1f43d9cfe782d782.js'},{'revision':null,'url':'/_next/static/chunks/app/auth/success/page-070ae95134c3f3e7.js'},{'revision':null,'url':'/_next/static/chunks/app/debug/page-8b906d41cb385b1c.js'},{'revision':null,'url':'/_next/static/chunks/app/install/page-4a619ab781fc3914.js'},{'revision':null,'url':'/_next/static/chunks/app/join/layout-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/app/join/page-ffd179b2863f3f66.js'},{'revision':null,'url':'/_next/static/chunks/app/layout-32c2bc6ee5f243ac.js'},{'revision':null,'url':'/_next/static/chunks/app/link/page-e1a00d7571240d6f.js'},{'revision':null,'url':'/_next/static/chunks/app/logout/page-1474173a72f77bf9.js'},{'revision':null,'url':'/_next/static/chunks/app/me/page-08898f94ae9de759.js'},{'revision':null,'url':'/_next/static/chunks/app/notifications/page-7ca63df0f65a3931.js'},{'revision':null,'url':'/_next/static/chunks/app/orders/page-8d083cd1200a5c81.js'},{'revision':null,'url':'/_next/static/chunks/app/overview/page-5147087d24205b8b.js'},{'revision':null,'url':'/_next/static/chunks/app/page-64056cb57b7df288.js'},{'revision':null,'url':'/_next/static/chunks/app/register/page-4403385d51642139.js'},{'revision':null,'url':'/_next/static/chunks/app/settlements/page-fe9be6e718a402b6.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/c/%5Bid%5D/page-59a4071a05136587.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/flash/page-c1b22a26fd53a49f.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/food-train/page-a0f77ec528b8b941.js'},{'revision':null,'url':'/_next/static/chunks/app/shop/page-8618309a1a9b73f2.js'},{'revision':null,'url':'/_next/static/chunks/app/spot/%5Bid%5D/page-af56c94bf6cf4248.js'},{'revision':null,'url':'/_next/static/chunks/app/spot/page-8357df49c76fe5c7.js'},{'revision':null,'url':'/_next/static/chunks/app/wallet/page-1655593dcfe0469d.js'},{'revision':null,'url':'/_next/static/chunks/framework-369218c9910949e7.js'},{'revision':null,'url':'/_next/static/chunks/main-app-b7c599722f79bdc1.js'},{'revision':null,'url':'/_next/static/chunks/main-be79bb95d5b42faf.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/app-error-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/forbidden-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/global-error-ee767baf43a5a2ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/not-found-27860661a41da3ba.js'},{'revision':null,'url':'/_next/static/chunks/next/dist/client/components/builtin/unauthorized-27860661a41da3ba.js'},{'revision':'846118c33b2c0e922d7b3a7676f81f6f','url':'/_next/static/chunks/polyfills-42372ed130431b0a.js'},{'revision':null,'url':'/_next/static/chunks/webpack-b695f13b2836b942.js'},{'revision':null,'url':'/_next/static/css/b71a7873b9c26fc2.css'},{'revision':'9771bcf20ba38f90712d273569b03e75','url':'/_next/static/lB-uHEsnE6PYLupb0Lu-R/_buildManifest.js'},{'revision':'b6652df95db52feb4daf4eca35380933','url':'/_next/static/lB-uHEsnE6PYLupb0Lu-R/_ssgManifest.js'},{'revision':'3fc49c2a6022fd7a30dd0fd643f57866','url':'/brand/banner.jpg'},{'revision':'27ce8d11af1fa5310a38e0a873b612c5','url':'/brand/logo.jpg'},{'revision':'21c8674ef5de87f15149b554bef954e8','url':'/icons/android/launchericon-144x144.png'},{'revision':'5a9f92af94660d63c0c436694774f624','url':'/icons/android/launchericon-192x192.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/android/launchericon-48x48.png'},{'revision':'5cc74a1180e9da94973b047042aded54','url':'/icons/android/launchericon-512x512.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/android/launchericon-72x72.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/android/launchericon-96x96.png'},{'revision':'7ee5446d081467e5800848c4ce3da6f3','url':'/icons/ios/100.png'},{'revision':'f11d312cd3d7cfee6f02a23906fd108f','url':'/icons/ios/1024.png'},{'revision':'9816127b8bdc3776952711b0977b2ef8','url':'/icons/ios/114.png'},{'revision':'0d05e20f509d2d38b37f0a3c101aa35d','url':'/icons/ios/120.png'},{'revision':'b98a05d66f7494066d0ba3a42aeca1ae','url':'/icons/ios/128.png'},{'revision':'2e0a67e030ee7061d57eb221fe780723','url':'/icons/ios/144.png'},{'revision':'e869e4705bd09e9b83a01afb1bd376d9','url':'/icons/ios/152.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/ios/16.png'},{'revision':'ea8d9c9d88c17d23c42850625c442fa0','url':'/icons/ios/167.png'},{'revision':'b6707a5ebfd9f53b3c6a9dc90a4082e0','url':'/icons/ios/180.png'},{'revision':'7fdcee2509439b092e668c9448cf9053','url':'/icons/ios/192.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/ios/20.png'},{'revision':'6d066b4ef4ee5b5dba9621aab8cc9a1a','url':'/icons/ios/256.png'},{'revision':'9005c4765e0f8e17d2eb63e028c24bf1','url':'/icons/ios/29.png'},{'revision':'19391d3de8599244346b47ad599d83e5','url':'/icons/ios/32.png'},{'revision':'0ce2683b3ee1866903723d275c6349fa','url':'/icons/ios/40.png'},{'revision':'629c5674e8fd89d33f57aaf092894285','url':'/icons/ios/50.png'},{'revision':'8cd35511a222d5cc4e4a23ef4c07688b','url':'/icons/ios/512.png'},{'revision':'34a8adcc3ef02c6cc74f86449481ac4d','url':'/icons/ios/57.png'},{'revision':'f9a363efb255f09f867333414e13458f','url':'/icons/ios/58.png'},{'revision':'eca69c492f9525f6661f80e844f89927','url':'/icons/ios/60.png'},{'revision':'324b63561172d305fc2f7021f71a527c','url':'/icons/ios/64.png'},{'revision':'2160847c205affc9d6c2850a9e67fdec','url':'/icons/ios/72.png'},{'revision':'b5c64f7e180c5b51e41deded738ed5ae','url':'/icons/ios/76.png'},{'revision':'254abe424ab5f8498480524eae0de96d','url':'/icons/ios/80.png'},{'revision':'e7fa488c29c40d108ec85e1f82a79d31','url':'/icons/ios/87.png'},{'revision':'84c7cad94d49288c8e93ba7e2679b33d','url':'/icons/windows/LargeTile.scale-100.png'},{'revision':'432a84dd02c63646e859edbb40df190c','url':'/icons/windows/LargeTile.scale-125.png'},{'revision':'096081a71cb71daa44c05e050695d8de','url':'/icons/windows/LargeTile.scale-150.png'},{'revision':'7608704a00592217d63f7b318c2e5f38','url':'/icons/windows/LargeTile.scale-200.png'},{'revision':'6e17ba59386ded6be1f10da2e5bfb9e0','url':'/icons/windows/LargeTile.scale-400.png'},{'revision':'3caa90776f06fc3ac443417ff6f9e9c2','url':'/icons/windows/SmallTile.scale-100.png'},{'revision':'267cadd8270e32c155b6dd3cd2af81b3','url':'/icons/windows/SmallTile.scale-125.png'},{'revision':'11c0fd9522841246d62af0fd09530459','url':'/icons/windows/SmallTile.scale-150.png'},{'revision':'d9596560b5df161ad68603a063323dea','url':'/icons/windows/SmallTile.scale-200.png'},{'revision':'fc07ec25a8ba517f6d69d36fc2bcfb5c','url':'/icons/windows/SmallTile.scale-400.png'},{'revision':'027040ff0058a27ef9bb9d957f0e1112','url':'/icons/windows/SplashScreen.scale-100.png'},{'revision':'1fc60c21b8e911b576605222e4604cac','url':'/icons/windows/SplashScreen.scale-125.png'},{'revision':'d7835db7d741404674b7e04de3648f24','url':'/icons/windows/SplashScreen.scale-150.png'},{'revision':'4069bf20884b973f3cb2256297917aae','url':'/icons/windows/SplashScreen.scale-200.png'},{'revision':'ac93274617295e12e608752ba70c9d5e','url':'/icons/windows/SplashScreen.scale-400.png'},{'revision':'4abc934b1324b56e1f66abbe6f2b2212','url':'/icons/windows/Square150x150Logo.scale-100.png'},{'revision':'121cd7794166b37d0b05f475eb55cc65','url':'/icons/windows/Square150x150Logo.scale-125.png'},{'revision':'39d68da5beb84c82745b7e562bbc6113','url':'/icons/windows/Square150x150Logo.scale-150.png'},{'revision':'31e6b17966d5c6da8ff8c6f0a78ea945','url':'/icons/windows/Square150x150Logo.scale-200.png'},{'revision':'4d68017168a3e5bb2173af09fcea4a6d','url':'/icons/windows/Square150x150Logo.scale-400.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.altform-lightunplated_targetsize-96.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.altform-unplated_targetsize-96.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.scale-100.png'},{'revision':'898a72b8a205ca8683f07e44931c5b3f','url':'/icons/windows/Square44x44Logo.scale-125.png'},{'revision':'c25a1b65e3a12b62d154556826ed1d16','url':'/icons/windows/Square44x44Logo.scale-150.png'},{'revision':'267cadd8270e32c155b6dd3cd2af81b3','url':'/icons/windows/Square44x44Logo.scale-200.png'},{'revision':'5044ecfe554402a9e9ef514a4ce20b3f','url':'/icons/windows/Square44x44Logo.scale-400.png'},{'revision':'02ee465259a5c774229d4c5dc7256a6b','url':'/icons/windows/Square44x44Logo.targetsize-16.png'},{'revision':'46b5bf291ba241e5cdb326e762c97b63','url':'/icons/windows/Square44x44Logo.targetsize-20.png'},{'revision':'ce0ebcaf38ea5138c1e4cd2d076f7ed3','url':'/icons/windows/Square44x44Logo.targetsize-24.png'},{'revision':'1ef77a059dba462f400143102e4d5ab7','url':'/icons/windows/Square44x44Logo.targetsize-256.png'},{'revision':'f917669a5e4bfc4b1f61fdec10cecc16','url':'/icons/windows/Square44x44Logo.targetsize-30.png'},{'revision':'410625ffcf66437d6049a15592af3430','url':'/icons/windows/Square44x44Logo.targetsize-32.png'},{'revision':'e17197251626e6ba820cdd02c96b6ad1','url':'/icons/windows/Square44x44Logo.targetsize-36.png'},{'revision':'b9b67d12a6c3fb93a439717e43a847a9','url':'/icons/windows/Square44x44Logo.targetsize-40.png'},{'revision':'66964a5c6a8537b83cb512660e79e17e','url':'/icons/windows/Square44x44Logo.targetsize-44.png'},{'revision':'c31ce7bd12408f071588d9da55c64c1f','url':'/icons/windows/Square44x44Logo.targetsize-48.png'},{'revision':'bdee1037809af22806052d53162d26c8','url':'/icons/windows/Square44x44Logo.targetsize-60.png'},{'revision':'e5a5ee6fb9be85bc51ad87416acc4991','url':'/icons/windows/Square44x44Logo.targetsize-64.png'},{'revision':'408da9e7542844421d76ede40f8191c4','url':'/icons/windows/Square44x44Logo.targetsize-72.png'},{'revision':'75e1945470ef164def8d220cb6fe31bc','url':'/icons/windows/Square44x44Logo.targetsize-80.png'},{'revision':'d69c14ae84ca1cf33ff79db600846a15','url':'/icons/windows/Square44x44Logo.targetsize-96.png'},{'revision':'06091ad8946496f71160e08f3acd694e','url':'/icons/windows/StoreLogo.scale-100.png'},{'revision':'47e34490c42fe1dae111b67301bf4b35','url':'/icons/windows/StoreLogo.scale-125.png'},{'revision':'bdd2d8c69c0be099276c4d6567fa5d9e','url':'/icons/windows/StoreLogo.scale-150.png'},{'revision':'111893dec64c28034160243677a0ffce','url':'/icons/windows/StoreLogo.scale-200.png'},{'revision':'9c66299aa54ba377775dc59fec049bde','url':'/icons/windows/StoreLogo.scale-400.png'},{'revision':'bc73480baeea24805571faccca15c354','url':'/icons/windows/Wide310x150Logo.scale-100.png'},{'revision':'77e3655ad0441ad2dd39395b2e641119','url':'/icons/windows/Wide310x150Logo.scale-125.png'},{'revision':'90b6c6c60ffa9081541da24cba21c3b3','url':'/icons/windows/Wide310x150Logo.scale-150.png'},{'revision':'027040ff0058a27ef9bb9d957f0e1112','url':'/icons/windows/Wide310x150Logo.scale-200.png'},{'revision':'4069bf20884b973f3cb2256297917aae','url':'/icons/windows/Wide310x150Logo.scale-400.png'},{'revision':'cc01aff3744d47049dea63726110b6e2','url':'/manifest.json'}],skipWaiting:!0,clientsClaim:!0,navigationPreload:!0,runtimeCaching:eq});self.addEventListener("push",e=>{let t=e.data?.json();if(!t)return;let a=t.title||"新訊息",s={body:t.body||"您有一則新通知",icon:"/icons/android/launchericon-192x192.png",badge:"/icons/android/launchericon-192x192.png",data:t.url||"/"};e.waitUntil((async()=>{await self.registration.showNotification(a,s);let e=await eC()+1;await eT(e),await eS(e)})())}),self.addEventListener("notificationclick",e=>{e.notification.close(),e.waitUntil(self.clients.openWindow(e.notification.data))}),eP.addEventListeners()})();
